[FEATURE] XML Sitemap
It is now possible to generate XML sitemaps for SEO purposes without
using 3rd-party plugins. When enabled, this new feature will create a
sitemapindex with one or more sitemaps in it. Out-of-the-box it will
have one sitemap containing all the pages of the current site and
language. Per site and per language you have the possibility to render
a different sitemap.

Resolves: #84525
Releases: master
Change-Id: Iad74b114b9dd37dbc4dd72e244437691fb8c31b5
Reviewed-on: https://review.typo3.org/58057
Reviewed-by: Björn Jacob <bjoern.jacob@tritum.de>
Reviewed-by: Richard Haeser <richard@maxserv.com>
Tested-by: Richard Haeser <richard@maxserv.com>
Tested-by: Kevin Appelt <kevin.appelt@icloud.com>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>