[TASK] Execute frontend functional tests in sub request
Constants TYPO3_MODE and TYPO3_REQUESTTYPE are unused
in core, and we have a new helper method in
typo3/testing-framework to execute sub requests.

The patch switches all functional tests that execute
frontend requests to use a sub request instead of
handling standalone PHP processes.

This patch is quite a milestone: It is the proof that
TYPO3 framework state can be managed well enough to
execute multiple TYPO3 application requests in one process.

As a happy little side effect, the overall performance
of the functional tests is increased by roughly 20
percent. This is quite a bit when considering the fact
that most execution time is consumed by setting up
all the database tables.

Releases: master
Resolves: #92966
Related: #92965
Related: #92947
Change-Id: I390baed39502ad3d30c881d6061f8ea0fad3c76c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65514
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>