[TASK] Update XLIFF version to 1.2
Update all XLIFF files to use version 1.2, which makes it possible to
use a validation schema.

Also drop the TYPO3-specific XLIFF file IDs (together with the reference
to the corresponding XML namespace) as those are not needed anymore,
but prevented XML validation.

Also drop checking the XLIFF file IDs from the CI jobs as this
does not make sense anymore now.

This change does not fix the discovered validation errors in the XLIFF
files. This will come later in a separate change.

Resolves: #93594
Releases: master, 10.4

Change-Id: I1f263b0a5a7f170e90a906d7992d7bbf08503f7d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/68676
Tested-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>