[BUGFIX] Render canonical link of target page when page is in mountpoint
When a page is within a mountpoint, the normal behaviour is to generate
a link within the mountpoint. As the content is shown from another page,
the url of the target page is necessary as a canonical to avoid
duplicate content. To do this, we temporarily remove the mountpoint
information while creating the canonical link which results in the
link to the target page.

Resolves: #93300
Releases: master, 10.4
Change-Id: Ie97c9e323d6faae5e928d6206d72a2082106c949
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/68469
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Franz Kugelmann <franz.kugelmann@elementare-teilchen.de>
Tested-by: Riny van Tiggelen <info@online-gamer.nl>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Richard Haeser <richard@richardhaeser.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Riny van Tiggelen <info@online-gamer.nl>
Reviewed-by: Richard Haeser <richard@richardhaeser.com>