[TASK] Avoid runtime overriding GLOBALS in functional tests
The InternalResponseContext class of the testing-framework
provides method withGlobalSettings() which can be used to
set arbitrary globals when executing a frontend sub request.

It is used in various tests to set some TYPO3_CONF_VARS.

The testing-framework already provides class property
$configurationToUseInTestInstance to set TYPO3_CONF_VARS
on a per test-case level.

The patch migrates tests towards this property and
avoids using an AdditionalConfiguration fixture file
that is also used to set some TYPO3_CONF_VARS.

Resolves: #96244
Releases: main, 11.5
Change-Id: Ie34a449075d1c1bacb1380dd5e79e3c48978bfe4
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72506
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72509
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>