[TASK] Align the test field annotations with the testing framework
Also drop type annotations that already are present in the
corresponding field in the superclass, and drop some type annotations
that are redundant with the existing native type declarations.

This helps resolves a bunch of new PHPStan warnings after the recent
changes in the testing framework. It also helps static code analysis,
and makes the code more succinct.

Resolves: #97687
Releases: main, 11.5
Change-Id: I6f1b67f410f7e195b036bc5b5b528cb271a990b9
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/74749
Tested-by: core-ci <typo3@b13.com>
Tested-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: crell <larry@garfieldtech.com>
Reviewed-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>