[BUGFIX] Make sure social media images gets processed if needed
The check if the image has to be processed has been moved to
dedicated methods to enhance readability and testability

The condition still checks if the image is wider than the defined
maximum width. The check if the entire image is used should now be
easier to understand. No offset and 100% of the width and height
of the image correspond to the full image and negated in its entirety.

Resolves: #98118
Releases: main, 11.5
Change-Id: I8283e2f74310654455fb247e84c9e4204f3e1dd4
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/75422
Tested-by: Sascha Egerer <sascha@sascha-egerer.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Benjamin Franzke <bfr@qbus.de>
Reviewed-by: Sascha Egerer <sascha@sascha-egerer.de>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>